feat: add Keyboard class
